refactor(webapp): inline the orb experiments and drop the logo cloud tab
The orbit-dots and logo-morph prototypes now live in the storybook route itself rather than a sibling module, keeping them contained to the experiment. Removes the logo cloud tab and the dot-cloud sampling it was the only user of.
